Redondeo números de los enunciados, Mostrar balance con espacios correctos, crea las tablas de un año pasado por parámetro

+++ src/es/ubu/inf/tfg/asientosContables/AportacionInicial.java
@@ -9,8 +9,8 @@ public class AportacionInicial extends Asiento {
 		fecha = f;
 		inputs = i;
 
-		String enunciado1 = " Cada uno de los " +inputs[0]+ " socios fundadores realiza una aportación inicial en "
-    			+ "efectivo a favor de la empresa por importe de " +inputs[1]+ "€ (" +(inputs[0])*(inputs[1])+ "€ en total). "
+		String enunciado1 = " Cada uno de los " +(int)inputs[0]+ " socios fundadores realiza una aportación inicial en "
+    			+ "efectivo a favor de la empresa por importe de " +inputs[1]+ "€ (" +inputs[0]*inputs[1]+ "€ en total). "
     			+ "Cada socio tiene una acción. La empresa ingresa el dinero recibido en su cuenta corriente. \n";
 		if (enunciadoCuentas){
 			enunciado1 = enunciado1 + "CUENTAS PGC: 100. Capital social; 572. Bancos e instituciones de crédito c/c vista, euros. \n";
